Fix : 서로 다른 건물이 동일한 위치에 건설되는 버그 수정

+++ Assets/IslandMonkey/Scripts/Fundamentals/HexagonalCalculator.cs
@@ -6,7 +6,7 @@ public class HexagonalCalculator
 	readonly Vector2 _qOffset;
 	readonly Vector2 _rOffset;
 
-	public HexagonalCalculator(int distance)
+	public HexagonalCalculator(float distance)
 	{
 		float xOffset = 0.5f * distance; // x
 		float yOffset = Mathf.Sqrt(3) * 0.5f * distance; // y
